add sample sets to the c prototype
We still gather the state for all samples, then union the state with the
sample set. This could improved on in the future, we could gather state
for only a subset of the samples, but for right now, it's a reasonable
approach given that the intermediate data does not consume that much
memory.