Fix issues with getting mutation samples
This is slightly less optimal, memory-wise, because now we have to store
the entire path from the mutation downward. We were only storing the
beginning and end of the path before, but it turns out if there's enough
complexity in the path, we can't rely on the first and last value to
represent the whole path. Now we have to store a 2d matrix of mutation
and node values that represents the path from the mutation to the bottom
of the tree. In my (not very quantitative) testing with the sgdp sample
set on chromosome 1, I'm not seeing much, if any, hits to memory or cpu
performance from this change.