Finish end to end tests, fix a few bugs
First, we need to move the result and its length outside of
process_tree, then we can test with python-generated test data. Since
the python prototype processes scenario 3 differently, we ignore that
one for now. Add test truth data.

Also, fix two small bugs. The first initializes the values in the
results array. This is necessary since we're adding to it and not ever
setting the first value directly. The second was a bug in the
normalization code. We were passing site_1 dimensions to the n_b
variable.

Add some comments along the way.